requirements.txt
flask
langchain==0.2.14
langchain-openai==0.1.7
python-dotenv
requests
gunicorn
